Use `contributors-short` macro to print `title-short`
One can gain more flexibility in citing abbreviated works in this way.
A citation in Markdown of the form '@author2000ShortTitle [27]' should
render something like 'Author (2000, 27)'. What Citeproc is doing in
that case is combining an 'author-only' citation with a 'supress-author'
citation. As a result, if you cite by title only (as 'classic' works are
cited on the current implementation of this style), a citation like
'@EN [1113a13]' is rendered as '(EE 1113a13)'. You can trick Citeproc by
giving the desired result ('EN (1113a13)') by leaving the 'author' field
blank for 'classic' entries. Because 'title' and 'title-short' are
fallbacks for the 'author' name variable, you can thereby get the title
to be what's included in an 'author-only' citation and the title to be
what's ommitted in a 'suppress-author' citation. This is an absolutely
ugly hack, but two low-level limitations stand in the way of anything
better. First is the fact that Citeproc only operates with these three
modes (the ones just mentioned, plus 'plain' citations). Second is the
fact that CSL citations are *always* enclosed by the designated
delimiters (in this case, '(' and ')'). As a result, to put it in
'biblatex' terms, with Citeproc and CSL it is only possible to generate
citations of the kinds '\textcite' and '\autocite', but not of the type
'\cite'.

Later, it might be worth whether a custom reader or writer can
accomplish anything that provides more flexibility. It might also be
possible to remove delimiters from the CSL style, and manually enclose
citations in parentheses. Then perhaps a Lua filter could strip the
parentheses when converting to LaTeX or when using a footnote-style
citation style when converting to other formats. In the meantime, this
is definitely a (surmountable) limitation on Markdown's usefulness for
writing more complex documents like articles.

Also simplify code excluding 'classic' works and works with
'title-short' from bibliography.